Add product details and styles
- Update  function to use brand for image paths
- Ensure images fit 518x518 dimensions in
- Add styles for main image and thumbnails
- Update HTML structure in
- Implement image change and animation in
