+++ src/pages/prodoucts-view.html
@@ -50,7 +50,6 @@
         </div>
         <!-- Here is where there are all the thumbnails -->
         <div class="thumbnail-container"></div>
-
       </div>
 
       <!-- Right Side: Product Details -->
